+++ packages/terra-date-picker/src/react-datepicker/calendar.jsx
@@ -7,7 +7,7 @@ import Month from './month'
 import React from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames/bind'
-import styles from './stylesheets/datepicker.scss'
+import styles from './stylesheets/datepicker.module.scss'
 import {
   now,
   setMonth,